+++ components/admin/ProjectDetailsForm/VideoProjectDetailsForm/VideoProjectDetailsForm.js
@@ -128,7 +128,6 @@ const VideoProjectDetailsForm = props => {
       visibility: {
         label: 'Visibility Setting',
         required: true,
-        props: { hide: 'internal' },
       },
       author: {
         label: 'Author',